listings/requirements.txt
django >= 3.2, < 3.3
djangorestframework >= 3.12, < 3.13
django-extensions==3.1.3
django-filter==2.4.0
psycopg2-binary==2.8.6